requirements/base.txt
amqp==2.6.1
arabic-reshaper==2.1.3
autopep8==1.5.7
billiard==3.6.4.0
celery==4.4.6
certifi==2021.5.30
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==1.11.17
django-allauth==0.34.0
django-celery-beat==1.1.1
django-celery-results==1.2.1
django-cors-headers==2.1.0
django-debug-toolbar==1.9
django-extensions==1.9.7
django-filter==2.1.0
django-model-utils==3.0.0
django-nested-admin==3.3.3
django-oauth-toolkit==1.0.0
django-pandas==0.6.4
django-redis==4.10.0
django-rest-auth==0.9.2
django-rest-swagger==2.1.2
django-silk==3.0.2
django-simple-history==1.9.0
django-stackoverflow-trace==0.0.4
djangorestframework==3.7.3
djangorestframework-filters==0.10.2
docutils==0.17.1
future==0.18.2
gevent==1.2.2
gprof2dot==2021.2.21
greenlet==1.1.2
gunicorn==19.7.1
html5lib==1.1
idna==2.6
importlib-metadata==4.8.1
itypes==1.2.0
Jinja2==3.0.1
kombu==4.6.11
lxml==4.6.3
MarkupSafe==2.0.1
model-mommy==1.4.0
mysqlclient==2.0.3
numpy==1.19.5
oauthlib==3.1.1
openapi-codec==1.3.2
pandas==1.1.5
Pillow==5.2.0
pycodestyle==2.7.0
pycpfcnpj==1.3
Pygments==2.10.0
PyJWT==2.1.0
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.1
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2017.3
raven==6.3.0
redis==3.5.3
reportlab==3.6.1
requests==2.18.4
requests-oauthlib==1.3.0
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
twilio==7.0.0
typing==3.7.4.3
typing-extensions==3.10.0.2
Unidecode==1.3.2
uritemplate==3.0.1
urllib3==1.22
vine==1.3.0
webencodings==0.5.1
Werkzeug==0.12.2
xhtml2pdf==0.2.5
zipp==3.6.0
